Worker pool with timeouts
In this version, all of the callers (in Node and in libuv)
call the TP with NULL timed_out and killed CBs.

However, the worker pool successfully times out the work in test/ehp/iodos-tp.js.
Huzzah!

NB This version of killed_cb assumes the caller's req is not destroyed earlier.
This assumption seems OK external callers using uv_queue_work,
but inappropriate for e.g. uv_fs_X wherein the req is managed by an external caller
and we want to invoke their done_cb with an error well before killed returns.

Need to revert to the version where the timed_out CB sets a data pointer
and we pass this along to the hangman to pass to the killed_cb.

Then the external party can set the pointer to the req itself if they want,
and the internal party can allocate a new structure on-demand for their resource
management policy.