get{addr,name}info: Safe behavior on timeout and killed
Although uv_getaddrinfo and uv_getnameinfo are read-only,
the calls to getaddrinfo and getnameinfo modify the memory to which you provide pointers.

In vanilla libuv, some of this memory is in a uv_req_t.
It may be deallocated after the return of uv__get{name,addr}info_done.

If a runaway get{name,addr}info syscall times out, it may continue to modify
the memory you gave it, even after uv__getXinfo_done returns.

Thus, we must give it memory that won't be cleaned up until the later of:
 - uv__getXinfo_done
 - uv__getXinfo_killed

and in uv__getXinfo_timed_out, we should give the TP the address of the buffer so it can be
passed to (and cleaned up in) uv__getXinfo_killed.