FIXUP: don't run prio done() synchronously after a timeout
Work is completed with a timeout. The worker won't be filing it so we must.
If w->prio, this is non-blocking, creating a helper thread to run done().
That way we can proceed to cancel'ing the thread and calling killed_cb.

This permits done() to synchronize with killed_cb, which may be necessary
for black-box libraries like in Node-C++ land.