fs: Make ReadDir Timeout-aware
1. node_watchdog: Add TimeoutPending for polling while Leashed

2. ReadDir is an unusual API.

When called synchronously, it calls readdir() using our sync2async technique.
However, it then mines the entire directory with a sequence of scandirs.

So, monitor for timeouts while we're doing the post-processing of scandir.
Otherwise the Event Loop can be poisoned by a giant directory.

While doing the post-processing, acquire a Leash.
Use the new LazyTimeoutWatchdog::TimedOut to poll the Watchdog to see if we're done.
This won't work for PreciseTimeoutWatchdog, ah well.

3. Allow internal::Isolate::Timeout to refuse to throw

Cowardly refusing to throw...

Here's the story:
- TimeoutWatchdog requests a TimeoutException be thrown.
- We throw a *new* TimeoutException, not a singleton like the TerminationException
- Thus there needs to be a HandleScope somewhere
- In (un-ported) C++ bindings there may not be a HandleScope

If no HandleScope, TimeoutWatchdog's Timeout() request causes
V8 to abort -- DoS!

In this situation, it's preferable to delay the TimeoutException
until we have a HandleScope.
That should occur on the next return to JavaScript,
at which point the Interrupt will be checked.

However, this condition really means you should port your C++ bindings.
I'm not sure which is preferable, TBH.

I believe we cannot delay into another callback (i.e. spurious timeouts).
We will reach a HandleScope before then.
All callbacks leave C++-land and finish in JS, at which point the TimeoutException can be delivered.